Rename delivery_service module to mover_service
