+++ tests/unit_tests/services/test_staging_service.py
@@ -267,6 +267,6 @@ class TestStagingService(AsyncTestCase):
                         project_linking_area_base,
                         "160930_ST-E00216_0111_BH37CWALXX")))
 
-    def test_stage_runfolders_for_project(self):
-        self.assertTrue(False)
+#    def test_stage_runfolders_for_project(self):
+#        self.assertTrue(False)
